Squashed some bugs... a few more to go Changes to be committed: 	modified:   algorithm1.cpp 	modified:   algorithm2.cpp 	modified:   main.cpp 	modified:   other.cpp 	modified:   shortest_path.cpp 	modified:   shortest_path.h

+++ shortest_path.h
@@ -5,7 +5,7 @@
 
 struct Vertex {
     int known;
-    int dist;
+    long int dist;
     int path;
 };
